Time: 48 ms (67.04%), Space: 30.1 MB (32.47%) - LeetHub

+++ 0115-distinct-subsequences/0115-distinct-subsequences.cpp
@@ -22,7 +22,7 @@ public:
             }
         }
         
-        return dp[n][m];
+        return (int)dp[n][m];
     }
     
     int f(int i, int j, string &s, string &t, vector<vector<int>>& dp)
